crypto/newhope: print values as unsigneds.
Otherwise builds fail with:
  crypto/newhope/newhope_statistical_test.cc:136:27: error: format specifies type 'long' but the argument has type 'uint64_t' (aka 'unsigned long long') [-Werror,-Wformat]

+++ crypto/newhope/newhope_statistical_test.cc
@@ -133,9 +133,9 @@ static bool TestKeys(void) {
   uint64_t bits = NEWHOPE_KEY_LENGTH * 8 * kNumTests;
   uint64_t diff = bits - 2 * ones;
   double fraction = (double) abs(diff) / bits;
-  printf("ones:   %ld\n", ones);
-  printf("zeroes: %ld\n", bits - ones);
-  printf("diff:   got %ld (%f), want %ld\n", diff, fraction, 0L);
+  printf("ones:   %u\n", (unsigned) ones);
+  printf("zeroes: %u\n", (unsigned) (bits - ones));
+  printf("diff:   got %u (%f), want %ld\n", (unsigned) diff, fraction, 0L);
   printf("\n");
 
   if (fraction > 0.01) {
